# openapi.yaml
openapi: 3.0.0
info:
  title: Tournaments API
  description: API used by the frontend
  version: 0.1.0

servers:
  - url: "http://localhost:8080/api/v1"
    description: local development testing

tags:
  - name: Tournaments
    description: Operations about tournaments
  - name: Series
    description: Operations about series
  - name: Games
    description: Operations about games
  - name: Users
    description: Operations about users
  - name: Teams
    description: Operations about teams

paths:
  /users:
    post:
      tags:
        - Users
      summary: Create user
      description: Create a user
      operationId: createUser
      requestBody:
        description: Create a new user
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: "#/components/schemas/User"
                required: ["user"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

    get:
      tags:
        - Users
      summary: Get users
      description: Returns a list of users.
      operationId: getUsers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
                required: ["users"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /teams:
    post:
      tags:
        - Teams
      summary: Create team
      description: Create a team
      operationId: createTeam
      requestBody:
        description: Create a new team
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Team"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  team:
                    $ref: "#/components/schemas/Team"
                required: ["team"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

    get:
      tags:
        - Teams
      summary: Get teams
      description: Returns a list of teams.
      operationId: getTeams
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  teams:
                    type: array
                    items:
                      $ref: "#/components/schemas/Team"
                required: ["teams"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /tournaments:
    get:
      tags:
        - Tournaments
      summary: Get tournaments
      description: Returns a list of tournaments.
      operationId: getTournaments
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  tournaments:
                    type: array
                    items:
                      $ref: "#/components/schemas/Tournament"
                required: ["tournaments"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /tournaments/{tournamentId}/series:
    get:
      tags:
        - Series
      summary: Get tournament series
      description: Returns a list of series for a tournament.
      operationId: getSeriesByTournamentId
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
              name: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    type: array
                    items:
                      $ref: "#/components/schemas/Series"
                required: ["series"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /tournaments/{tournamentId}/games:
    get:
      tags:
        - Games
      summary: Get games
      description: Returns a list of games for a tournament.
      operationId: getGamesByTournamentId
      parameters:
        - name: tournamentId
          in: path
          description: ID of tournament
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
              name: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  games:
                    type: array
                    items:
                      $ref: "#/components/schemas/Game"
                required: ["games"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /series:
    post:
      tags:
        - Series
      summary: Create series
      description: Create a series
      operationId: createSeries
      requestBody:
        description: Create a new series
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Series"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    $ref: "#/components/schemas/Series"
                required: ["series"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /series/{seriesId}:
    get:
      tags:
        - Series
      summary: Get series
      description: Returns a list of series by ID.
      operationId: getSeriesById
      parameters:
        - name: seriesId
          in: path
          description: ID of series
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
              name: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    $ref: "#/components/schemas/Series"
                  games:
                    type: array
                    items:
                      $ref: "#/components/schemas/Game"
                required: ["series", "games"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags:
        - Series
      summary: Update series
      description: Update an existing series
      operationId: updateSeriesById
      parameters:
        - name: seriesId
          in: path
          description: ID of series
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
              name: uuid
      requestBody:
        description: Update an existing series
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Series"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  series:
                    $ref: "#/components/schemas/Series"
                required: ["series"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
  /games:
    post:
      tags:
        - Games
      summary: Create game
      description: Create a game
      operationId: createGame
      requestBody:
        description: Create a new game
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: "#/components/schemas/Game"
                required: ["game"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

  /games/{gameId}:
    get:
      tags:
        - Games
      summary: Get games
      description: Returns a game by ID.
      operationId: getGameById
      parameters:
        - name: gameId
          in: path
          description: ID of game
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
              name: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: "#/components/schemas/Game"
                required: ["game"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"
    put:
      tags:
        - Games
      summary: Update an existing game
      description: Update an existing game
      operationId: updateGameById
      parameters:
        - name: gameId
          in: path
          description: ID of game
          required: true
          schema:
            type: string
            format: uuid
            x-go-type: uuid.UUID
            x-go-type-import:
              path: github.com/google/uuid
              name: uuid
      requestBody:
        description: Update an existing game
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Game"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  game:
                    $ref: "#/components/schemas/Game"
                required: ["game"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/ServerError"

components:
  schemas:
    Error:
      type: object
      properties:
        code:
          type: integer
        message:
          type: string
      required:
        - code
        - message
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: primarykey
        name:
          type: string
        shortname:
          type: string
      required:
        - id
        - name
        - shortname
    Team:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: primarykey
        name:
          type: string
        slug:
          type: string
      required:
        - id
        - name
        - slug
    Tournament:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: primarykey
        name:
          type: string
      required:
        - id
        - name
    Series:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: primarykey
        tournamentId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: index
        round:
          description: round number that this series is a part of (0, 1, 2)
          type: integer
        firstTeamId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        secondTeamId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        firstTeamUserId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        secondTeamUserId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        firstTeamScore:
          type: integer
        secondTeamScore:
          type: integer
      required:
        - id
        - tournamentId
        - round
    Game:
      type: object
      properties:
        id:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: primarykey
        seriesId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-oapi-codegen-extra-tags:
            gorm: index
        homeTeamId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        awayTeamId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        homeTeamUserId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        awayTeamUserId:
          type: string
          format: uuid
          x-go-type: uuid.UUID
          x-go-type-import:
            path: github.com/google/uuid
            name: uuid
          x-go-type-skip-optional-pointer: true
        homeTeamScore:
          type: integer
        awayTeamScore:
          type: integer
      required:
        - id
        - tournamentId
        - seriesId
  responses:
    BadRequest:
      description: A client side bad request error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ServerError:
      description: A server error occurred
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"